container: split into reader+writer
